Removed the IMutable interface which propagated notifications about value changes.
It do not provide 100% guarantee of notification.  For example:
 - If the user obtained a reference to same value from two places (eg classA.someInt and classB.someInt where classA == classB) then change in one place did not trigger notification in the other.  
 - Changing of the backing value for a property would not change the property.

Idea for better system (if it is needed in the future):
Each value refers to some memory range (eg reference it is 4 byte pointer in memory;  string is value on heap).
Whenever a debugger is resumed all bets are off and anything could have changed.  
Other then that value changes if and only if its memory range is modified by the user.
Expression as a whole changes whenever any of the values involved changes;  if method call or property is involved then the whole expression may change when any memory is modified.

git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@2772 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61