Object graph visualizer - replaced Graphviz's neato.exe by my custom spline routing. Removed all native Graphviz libraries.
Routing algorithm (Graphviz's algorithm is similar): search shortest path (Dijkstra - fast enough, A* would be faster) in the following graph: vertices are corners off all boxes, moved outside of the boxes a little. Edges: any pair of vertices that can be connected by straight line not while not intersecting any box. Having the shortest path as poly-line. Just curve the corners (looks surprisingly good). Edges are routed independently, so they can cross. Having our own implementation, we could work on edge non-crossing algorithm in the future. That is, if the whole layout and routing won't be replaced by GraphSharp in the end.

git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@5665 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61