First implementation of GUI for editor context actions.
A WPF Popup is shown over the editor when some actions are available. Can be opened by mouse click (TODO open by shortcut).
A Popup is probably not the best solution: it is not "tied" to current editor line anyhow so it stays open when region is collapsed, active ViewContent changes etc. (for active ViewContent change I added explicit handling). It should be probably part of editor's visual tree, in a special layer. Alternatively it could be displayed in the bookmark margin, but that would make it too small. This way it resembles ReSharper.
Have to consult this with Daniel.

git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@6290 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61