Dispose ProjectChangeWatcher when solution is disposed. Fixed adding 'AnyCPU' platform to a single project using the configuration editor.
Save "GlobalSection(NestedProjects)" in the correct order - SharpDevelop was saving it in the opposite order of the "Project"-sections. This was the cause of the whole .sln-file getting rearranged on every change, making merging .sln changes difficult.
Note that fixed SharpDevelop version will rearrange the solution one last time if the "NestedProjects" global section specifies a different order than the "Project"-sections.