BuildWorker: swap communication and build threads.
This is an attempt to fix http://community.sharpdevelop.net/forums/t/15654.aspx - BuildWorker40 hangs on windows 7 x64bit when invoking the markup compiler.
According the the report in the forums, the problem disappears when MSBuild is invoked on the main thread, and communication is suspended while the build is in progress (breaking cancellation).
This patch also moves the communication onto a background thread, fixing build cancellation. This might solve the problem if the hang was somehow related to the thread MSBuild was running on. However, if the AppDomain unload fails due to the thread stuck reading from the named pipe, the build worker might continue hanging with this patch.