[WIP] Major refactoring of the unit testing addin:
A lot of functionality gets moved from the main portion of the AddIn into the framework-specific providers.

We no longer deal with test classes and members, but instead use the new ITest interface.
A test can be a whole project, namespace, class, or just a single method.
This can be used to support features like the NUnit [TestCase] attribute, where a single method might result in multiple tests.