Fix null reference exception when running project.
BeforeBuildCustomToolProjectItems was not handling IBuildable derived classes that were not either a Project or Solution. This happens when a project is run and SharpDevelop is configured to build modified projects or build modified projects and dependent projects or when running unit tests.
Now custom tools are only run when a project or solution is explicitly rebuilt or built using the Build/Rebuild from the Build menu.