Fix changing Source Analysis project options not showing as dirty.
Problem occurred if the project had no SourceAnalysisOverrideSettingsFile set in the project.
In this case the options panel was overriding the ProjectOptions.Load method and changing the
values but not setting the IsDirty flag back to false so the UI was showing it as not dirty
when in fact it had been changed. Further changes did not cause the IsDirty flag to change so
the UI never displayed the an asterisk indicating there were outstanding changes.