Fixes bug #10731
https://bugzilla.xamarin.com/show_bug.cgi?id=10731

It will look as a separate issue checker for basic interfaces and it will give the first implementation that works as base (so it will not give 5 possible conversions, even all are possible). In practice works really well. It is losely based on the ParameterCanbeDemotedIssue so I exposed the code to not duplicate it. Yet the code is much simpler this action.