Added default ReportErrors delegate that does nothing.
This is necessary after the fix in commit 295f62532f3be6cb42ed15128806afe08ca67a10.
Before that commit the ReportErrors was semi-optional; When loading was successful ReportErrors was never called (ReportErrors optional), but when document failed to be created a NullReferenceException was thrown because ReportErrors did not have a delegate.

See also commit 006558f253769244bfd0141b75d8ce1a6c52b62c, that throws a more friendly exception than NullReferenceException when document fails to load.