Fix incorrect package sources being used when updating a package.
On updating a package all the enabled package sources are used. On
changing the enabled package sources, after opening a solution, the
original set of enabled package sources were being still used.

The change to support solution specific NuGet.Config files broke the
original behaviour since the package sources are reloaded when a
solution is opened or closed and the package repository cache was
still using the original set of package sources.