Generate repositories.config file after NuGet package restore.
The repositories.config contains a list of projects that share the
packages directory. If this is not restored then uninstalling a NuGet
package from one project could end up with NuGet removing it
completely from the packages directory since it believes that no other
project is referencing this NuGet package even if this is not the case.