Initial port to NuGet 3.0
This is an early version that is very broken.

Uses a custom build of NuGet 3.0 which targets .NET 4.5:

https://github.com/mrward/nuget/tree/3.0-ctp1-net45

Official binaries target .NET 4.5.1 so they cannot be used with
SharpDevelop.

NuGet binaries are not added to source control since this branch
is still a work in progress. So NuGet will need to be compiled from
source and the binaries copied to the RequiredLibraries directory.

It looks like SharpDevelop may get the automatic rollback that
happens in Visual Studio when a package fails to install since it
looks like the code is now in NuGet.Core.

What works:

Installing a package.
Uninstalling a package.
Updating a package.

Issues:

1. Uninstalling/updating does not remove the old files from the
packages directory.
2. Managing at the solution level is not implemented.
3. Fails to connect to new NuGet 3.0 feed. Getting an Http version not
supported error back from the NuGet 3.0 package source.
4. Unit tests do not compile due to various API changes.