Fix installing, uninstalling and updating solution level packages.
Can now install, uninstall and update a package such as OpenCover
which is installed in the solution but not in any project.

PackageViewModel class needs to be refactored to remove the code
duplication.

Also NuGet creates the .nuget/NuGet.config file automatically for the
solution level package. However it does not remove it when the
solution level package is uninstalled which may be a problem with the
NuGet addin.