Use version constraints for updates shown in Manage Packages dialog.
A version constraint defined in the packages.config file, as shown
below, is now used when checking for package updates from a package
source. Previously always the latest version of the NuGet package
would be displayed as an update in the Manage Packages dialog.

  <package
    id="jQuery"
    version="1.4.1"
    targetFramework="net40"
    allowedVersions="[1.4.1,1.8)" />

After updating to jQuery 1.7.2 then no further updates will be shown
in the Manage Packages dialog.

Currently this only works if a project is selected when the Manage
Packages dialog is opened.