Fix #563 null reference exception when managing packages.
A null reference exception would occur if the Manage Packages dialog
was opened when the solution was selected, the solution contained
only one project, and then a package was installed. If the solution
had more than one project the exception would not occur.

The NuGet addin was incorrectly treating a solution with one project
the same as if the project was selected when the Manage Packages
dialog was opened.