Handle any exceptions when NuGet package added to project.
Logic to check if the NuGet package is for the solution or for the
project was not handling any exceptions that could be generated.
For example if installing a NuGet package that supports the
dnx framework, which is not currently supported in NuGet 2.8.3,
the exception would not be caught because the try catch was at
a lower level and not at the top level.

Exceptions are now caught when checking the package to see if it is a
solution level package when adding, removing or managing a NuGet
package.