Open readme.txt when NuGet package is installed.
SharpDevelop will now open the readme.txt file in a NuGet package when
it is explicitly installed or updated. The readme.txt file needs to be
in the root directory of the package. If the NuGet package is being
installed as a dependency of another NuGet package then the readme.txt
will not be opened.