Show package not found error message for Update-Package -reinstall
Fixed null reference exception when package not found when
reinstalling a package for a single project.

Throw an exception when the package id cannot be found when
reinstalling a package for a single project or if the package id
cannot be found in any project when reinstalling for all projects.