Handle error looking for init.ps1 in NuGet packages directory.
If there is an exception finding init.ps1 files when the solution is
opened then the unhandled exception dialog is shown. This can happen
if a NuGet package requires NuGet 3.0. This results in an
NuGetVersionNotSatisfiedException being thrown. Now when this happens
the error message is reported in the PowerShell console if it is open.