Fix properties added by NuGet package PowerShell script being encoded
The SlowCheetah NuGet package uses the MSBuild API to directly modify
the MSBuild project file stored in memory in MSBuild's
GlobalProjectCollection. It adds properties that contain special
MSBuild characters, for example $( ). When SharpDevelop was merging
these new properties back into the project file on disk it was using
MSBuild to encode them so the wrong values were added. Now the
property values are not encoded.